Update styling and content for better visual hierarchy and accuracy
The changes in this commit are primarily focusing on updating the hierarchical visual appearance of the pages along with some text update. The H1, H2 and H3 styles have been decreased in size to improve the visual hierarchy. A minor padding adjustment was also added to the div in current-projects/page.tsx to enhance its styling. Some text content was removed from new-textbooks/page.tsx and center-library/page.tsx to streamline information while in employees/page.tsx, some of the employee designations were corrected as per latest data. These changes help to improve the readability and accuracy of information throughout the app.